requirements.txt
flask
python-dotenv
apscheduler
landingai-ade
pypdf
pydantic
psycopg[binary]
stripe
google-generativeai
anthropic
